requirements.txt
numpy>=2.0
python-dotenv
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-vector-stores-qdrant
qdrant-client>=1.9.0
openai>=1.0.0
pdfplumber
pymupdf
pandas
openpyxl
pytesseract
Pillow
python-pptx
python-docx
requests
 